backend/requirements.txt
django
psycopg2
django-health-check 
django-cors-headers
requests
djangorestframework
djangorestframework-simplejwt
